refactor(ui): flatten `cluster-workflow-templates/components/` directory
- some components were in the `components/` dir and others were nested
  - a folder for a single file is unnecessary and makes reading harder
  - a folder for two files, an SCSS file and the TSX file, is not much better either
    - they appear right next to each other in editors as they're alphabetized by filename anyway, and the file ending is last
    - and in some editors/themes, are clearly distinguished by icons as well
