+++ ui/src/app/cluster-workflow-templates/components/cluster-workflow-template-container.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import {Route, RouteComponentProps, Switch} from 'react-router';
-import {ClusterWorkflowTemplateDetails} from './cluster-workflow-template-details/cluster-workflow-template-details';
-import {ClusterWorkflowTemplateList} from './cluster-workflow-template-list/cluster-workflow-template-list';
+import {ClusterWorkflowTemplateDetails} from './cluster-workflow-template-details';
+import {ClusterWorkflowTemplateList} from './cluster-workflow-template-list';
 
 export const ClusterWorkflowTemplateContainer = (props: RouteComponentProps<any>) => (
     <Switch>
