+++ ui/src/app/cluster-workflow-templates/components/cluster-workflow-template-list.scss
